Add helper for running smoke tests
